// src/app/_Models/company-dto.ts
export class CompanyDTO {
    CompanyId: number
    CompanyName: string
    Address: string
    Phone: string
    CountryCode:string;
    Email: string
    Fax: string
    CountryId: number;
    CityId: number;
    Country: string;
    City: string
    OrganizationId: number
    IsActive: boolean
    latitude: string
    longitude: string
    Code: string
    CompanyEmail: string
    message: number

    CreatedBy: number
    FlagId: number
    RoleId: number;
    PageNumber:number;
    PageSize:number;
    Search:string;
    Quotes:string;
}